Backend/requirements.txt
Django>=4.2
python-dotenv
django-cors-headers
djangorestframework
pymongo
dnspython
bcrypt
razorpay
PyJWT
Pillow>=10.0.0